Add baseline and startup profiles.
