+++ android/benchmark/src/main/java/io/github/reactivecircus/benchmark/utils.kt
@@ -1,0 +1,5 @@
+package io.github.reactivecircus.benchmark
+
+import androidx.test.platform.app.InstrumentationRegistry
+
+val PackageName: String = InstrumentationRegistry.getArguments().getString("targetAppId")!!
